chore: Update MarkdownLatexParser to support LaTeX math equations
