+++ frontend/src/components/layout/Conversation/ConversationBody/MessageItem.jsx
@@ -1,20 +1,20 @@
-import React, { useState, useCallback, useEffect, useMemo, useRef } from 'react';
+import React, { useState, useCallback, useEffect, useRef } from 'react';
 import TextareaAutosize from 'react-textarea-autosize';
 import PropTypes from 'prop-types';
 import { useTranslation } from 'react-i18next';
 
 import ModalDialog from '../../../common/ModalDialog/ModalDialog';
 import { messageShape, conversationShape } from '../../../../model/conversationPropType';
-import { convertTextToSpeech } from '../../../../api/textToSpeechModelService';
 import { handleKeyDown as handleKeyDownUtility } from "../../../common/util/useTextareaKeyHandlers";
 import { userShape } from "../../../../model/userPropType";
 import FileItem from '../ConversationFooter/FileItem';
 import { deleteFile } from "../../../../api/fileService";
 import MarkdownLatexParser from './MarkdownLatexParser';
+import useVoice from './UseVoice';
+
 import './MessageItem.scss';
 
 const maxLineCount = 4;
-const pauseDuration = 600;
 
 function MessageItem({ message, onDelete, onEdit, user, setError, currentConversation, setCurrentConversation }) {
   const { t } = useTranslation();
@@ -23,103 +23,12 @@ function MessageItem({ message, onDelete, onEdit, user, setError, currentConvers
   const [isEditing, setIsEditing] = useState(false);
   const [editedMessage, setEditedMessage] = useState(message.content);
   const [isExpanded, setIsExpanded] = useState(true);
-  const [isSpeaking, setIsSpeaking] = useState(false);
   const textareaRef = useRef(null);
-  const audioRef = useRef(null);
   const resetCopyState = useCallback(() => {
     setCopied(false);
   }, []);
 
-  function getVoice(currentConversation, message, userSettings) {
-    if (currentConversation.isAIConversation && currentConversation.voice1 && currentConversation.voice2) {
-      const index = currentConversation.messages.findIndex(mess => mess.messageId === message.messageId);
-      if (index === -1) {
-        return userSettings.textToSpeechModel.voice_id;
-      }
-      return index % 2 === 0 ? currentConversation.voice1 : currentConversation.voice2;
-    }
-    return userSettings.textToSpeechModel.voice_id;
-  }
-
-  const voice = getVoice(currentConversation, message, user.settings);
-  const textToSpeechVendor = (currentConversation.isAIConversation ? currentConversation.textToSpeechVendor : user.settings.textToSpeechModel.vendor) ?? user.settings.textToSpeechModel.vendor;  
-  const textToSpeechModelId = (currentConversation.isAIConversation ? currentConversation.textToSpeechModelId : user.settings.textToSpeechModel.model_id) ?? user.settings.textToSpeechModel.model_id;    const handleSpeakClick = useCallback(async () => {
-    if (isSpeaking) {
-      if (audioRef.current) {
-        audioRef.current.pause();
-        audioRef.current.currentTime = 0;
-      }
-      setIsSpeaking(false);
-      return;
-    }
-
-    setError(null);
-    setIsSpeaking(true);
-    const textChunks = message.content.split('\n').filter(chunk => chunk.trim() !== '');
-
-    // Function to convert text to speech and return a promise that resolves to the audio blob
-    const convertChunkToSpeech = (chunk) => {
-      const speech = convertTextToSpeech(
-        textToSpeechModelId,
-        chunk,
-        voice,
-        textToSpeechVendor
-      );
-      return speech;
-    };
-
-    // Function to play the audio blob
-    const playAudioBlob = async (audioBlob) => {
-      const audioUrl = URL.createObjectURL(audioBlob);
-      if (audioRef.current) {
-        URL.revokeObjectURL(audioRef.current.src); // Clean up previous audio object URL
-      }
-      audioRef.current = new Audio(audioUrl);
-
-      return new Promise((resolve) => {
-        audioRef.current.onended = () => resolve();
-        audioRef.current.play().catch(err => console.error('Playback error:', err));
-      });
-    };
-
-    // Asynchronously prefetch and play audio chunks
-    const prefetchAndPlayChunks = async () => {
-      try {
-        if (textChunks.length === 0) {
-          throw new Error('No text chunks to process');
-        }
-
-        let hasNextChunk = true;
-        let nextChunkIndex = 0;
-        let nextAudioBlobPromise = convertChunkToSpeech(textChunks[nextChunkIndex]);
-
-        while (hasNextChunk) {
-          const currentAudioBlobPromise = nextAudioBlobPromise;
-
-          nextChunkIndex++;
-          hasNextChunk = nextChunkIndex < textChunks.length;
-          nextAudioBlobPromise = hasNextChunk ? convertChunkToSpeech(textChunks[nextChunkIndex]) : null;
-
-          const audioBlob = await currentAudioBlobPromise; // Wait for the current audio blob
-          await playAudioBlob(audioBlob); // Play current chunk
-
-          // Wait for the specified pause duration before proceeding, unless it's the last chunk
-          if (hasNextChunk) {
-            await new Promise(resolve => setTimeout(resolve, pauseDuration));
-          }
-        }
-      } catch (error) {
-        setError(error.message); // Use setError here to handle the error
-      } finally {
-        setIsSpeaking(false); // Reset state when all chunks have been played or an error occurs
-      }
-    };
-
-    prefetchAndPlayChunks().catch(error => {
-      console.error('Error processing text to speech:', error);
-      setIsSpeaking(false);
-    });
-  }, [isSpeaking, message.content, setError, textToSpeechModelId, textToSpeechVendor, voice]);
+  const { isSpeaking, handleSpeakClick } = useVoice(user, currentConversation, message, setError);
 
   const handleModalClose = useCallback(() => setIsModalOpen(false), []);
 
@@ -206,19 +115,7 @@ function MessageItem({ message, onDelete, onEdit, user, setError, currentConvers
       console.error("Failed to delete file:", error);
     }
   }, [currentConversation.userId, message.messageId, setCurrentConversation]);
-  const processedContent = useMemo(() => {
-    return convertDelimiters(message.content.replace(/(?<!\n)\n(?!\n)/g, '  \n'));
-  }, [message.content]);
 
-  function convertDelimiters(text) {  
-    // Convert block delimiters \[...\] to $$...$$  
-    const blockRegex = /\\\[(.*?)\\\]/gs;  
-    const convertedBlockText = text.replace(blockRegex, (match, innerContent) => `$$${innerContent}$$`);
-  
-    // Convert inline delimiters \(...\) to $...$  
-    const inlineRegex = /\\\((.*?)\\\)/gs;  
-    return convertedBlockText.replace(inlineRegex, (match, innerContent) => `$${innerContent}$`);  
-  }
   return (
     <>
       {isModalOpen && (
@@ -310,7 +207,7 @@ function MessageItem({ message, onDelete, onEdit, user, setError, currentConvers
               ref={textareaRef}
             />
           ) : (
-            <MarkdownLatexParser content={processedContent} />
+            <MarkdownLatexParser content={message.content} />
           )}
         </div>
       </div>
